Move Viewfinder into UI Folder
