+++ app/src/main/java/dev/digitaldreamweavers/qrguard/ui/ViewFinderViewModel.java
@@ -1,4 +1,4 @@
-package dev.digitaldreamweavers.qrguard;
+package dev.digitaldreamweavers.qrguard.ui;
 
 import androidx.lifecycle.ViewModel;
