+++ app/src/main/res/layout/fragment_view_finder.xml
@@ -3,11 +3,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewFinderFragment">
+    tools:context=".ui.ViewFinderFragment">
 
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="Hello" />
+        />
 
 </FrameLayout>
